refactor: split onboard tail support
